#18 login logout

+++ src/Controller/Admin/AdminController.php
@@ -3,13 +3,13 @@
 namespace App\Controller\Admin;
 
 use App\Entity\Post;
+use App\Entity\User;
 use App\Repository\PostRepository;
 use App\Repository\CommentRepository;
 use App\Controller\AbstractController;
 
 Class AdminController extends AbstractController
 {
-
     public function index() 
     {
         $posts = $this->postRepository->findAll();
@@ -18,20 +18,47 @@ Class AdminController extends AbstractController
         ]);
     }
 
+    public function login()
+    {
+        if ($this->getCurrentUser() and $this->getCurrentUser()->getIsAdmin()) {
+            return $this->index();
+        } 
+
+        if (isset($_POST['submit'])) {
+            $user = $this->userRepository->findOneByEmail($_POST['email']);
+
+            if (password_verify($_POST['password'], $user->getPassword())) {
+                $_SESSION['user'] = $user;
+
+                if ($_SESSION['user']->getIsAdmin()) {
+                    Header('Location: /admin/posts');
+                } else {
+                    var_dump('Vous devez être administrateur pour entrer ici!'); exit;
+                }
+            } else {
+                var_dump('Le mot de passe est invalide.'); exit;
+            }
+        }
+
+        return $this->twig->render('admin/login.html.twig');
+    }
+
     public function addPost()
     {
         if (isset($_POST['submit'])) {
-            $post = new Post(
-                $_POST['title'],
-                $_POST['intro'],
-                $_POST['content'],
-                $_POST['image']
-            );
-            $postId = $this->postRepository->insert($post);
-            
-            header("Location:/admin/edit/post/$postId");
+            if ($this->getCurrentUser() and $this->getCurrentUser()->getIsAdmin()) {
+                $post = new Post(
+                    $_POST['title'],
+                    $_POST['intro'],
+                    $_POST['content'],
+                    $_POST['image'],
+                    $_SESSION['user']
+                );
+                $postId = $this->postRepository->insert($post);
+                
+                header("Location:/admin/edit/post/$postId");
+            }
         }
-        
         return $this->twig->render('admin/add.html.twig');
     }
 
@@ -40,16 +67,22 @@ Class AdminController extends AbstractController
         $post = $this->postRepository->findOneById($id);
 
         if (isset($_POST['submit'])) {
-            $post->setTitle($_POST['title']);
-            $post->setIntro($_POST['intro']);
-            $post->setContent($_POST['content']);
-            $post->setImageUrl($_POST['image']);
+            $current_user = $this->getCurrentUser();
 
-            $this->postRepository->edit($post);
+            if ($current_user 
+            and $current_user->getIsAdmin()
+            and $current_user->getId() === $post->getAuthor()->getId()) 
+            {
+                $post->setTitle($_POST['title']);
+                $post->setIntro($_POST['intro']);
+                $post->setContent($_POST['content']);
+                $post->setImageUrl($_POST['image']);
 
-            header("Location:/admin/posts");
+                $this->postRepository->edit($post);
+
+                header('Location: /admin/posts');
+            }
         }
-
         return $this->twig->render('admin/edit.html.twig', [
             'post' => $post
         ]);
@@ -57,13 +90,21 @@ Class AdminController extends AbstractController
 
     public function deletePost(int $id)
     {
-        if (isset($_POST['delete'])) {
+        if (isset($_POST['delete'])) 
+        {
+            $current_user = $this->getCurrentUser();
             $post = $this->postRepository->findOneById($id);
+
+            if ($current_user 
+            and $current_user->getIsAdmin()
+            and $current_user->getId() === $post->getAuthor()->getId()) 
+            {
+
             $this->postRepository->delete($post);   
 
             header('Location:/admin/posts');
+            }
         }
-        
     }
     
     public function showCommentsFromPost(int $id)
@@ -81,7 +122,8 @@ Class AdminController extends AbstractController
 
     public function deleteComment(int $id)
     {
-        if (isset($_POST['delete'])) {
+        if (isset($_POST['delete'])) 
+        {
             $comment = $this->commentRepository->findOneById($id);
             $this->commentRepository->delete($comment); 
 
@@ -91,7 +133,8 @@ Class AdminController extends AbstractController
 
     public function approveComment(int $id, bool $validate)
     {
-        if (isset($_POST['unvalidate']) || isset($_POST['approve'])) {
+        if (isset($_POST['unvalidate']) || isset($_POST['approve'])) 
+        {
             $comment = $this->commentRepository->findOneById($id);
             $comment->setIsValidated($validate); 
             $this->commentRepository->approve($comment);
